GraspNeRF-main/requirements.txt
wheel
torch==1.10.0+cu111
torchvision==0.11.1+cu111
torchaudio==0.10.0+cu111
pybullet==2.7.9
pytorch-lightning==1.5
spconv-cu111==2.1.21
scikit-image==0.18.0
pyglet==1.5.15
easydict
plyfile
numpy==1.23.4
pyyaml
h5py
inplace-abn
opencv-python
tqdm
matplotlib
scipy
lpips
transforms3d
kornia
scikit-learn
catkin_pkg
black
jupyterlab
pandas
mpi4py
open3d
pytorch-ignite
tensorboard
numba
pillow
wandb
imageio
numpy_indexed
pyrender
ray
trimesh
pycuda
urdfpy
cython
monai